Revert "[fix] 팔로워, dm수정 "

+++ src/main/java/com/mopl/api/domain/conversation/dto/response/conversation/ConversationResponseDto.java
@@ -11,7 +11,7 @@ public record ConversationResponseDto(
     String nextCursor,
     UUID nextIdAfter,
     boolean hasNext,
-    Long totalCount,
+    int totalCount,
     String sortBy,
     String sortDirection
 ) {
