+++ src/main/java/com/mopl/api/domain/conversation/dto/response/direct/DirectMessageResponseDto.java
@@ -12,7 +12,7 @@ public record DirectMessageResponseDto(
     String nextCursor,
     UUID nextIdAfter,
     boolean hasNext,
-    Long totalCount,
+    int totalCount,
     String sortBy,
     String sortDirection
